Add Address Identity parameter to simple search
In some cases, an address ID should be able to be used to return a set
of metadata. To prevent duplication of code, and messing with the
chevalier threading mvar, I've added an "address" field to the simple
search.

If address is supplied, it's added to the protobuf. If it is not
supplied, then it is filled with nothing.

Limitations:
 - fromMaybe used to denote "no value" in SimpleSearch is used as a case
   in Chevalier to check if it defaulted out. This is suboptimal.
 - searching can be invoked without either an address nor query, or one
   or the other. In the case of both terms, the Query is not recognised
by Chevalier, however, I cannot workout how to build an empty SourceTags
to suit this upstream behaviour.