Merge key/value searching into formatted 'q' queries
Logic: q queries can be of the form `?q=string` or `?q=key:value`, with
a
colon delimiter. If there is a colon, the value is split into `("key",
"value")` and passed directly to chevalier for searching. Any wildcards
must be provided by the client. If there is no colon, it is passed in
the form `("*", "*" + string + "*")` - a wildcard key, and wildcard
string
wrapping.

Exact searches for values with no known key is achieved by using the
form `?q=*:value`, which will be passed as `("*","value")`.